Started to build GameLogic
gameWon and gameOver can both be set to false initially and changed to true later when conditions are met in the main method but gameLegal is conditional on the input and current state of the board. Need to decide first how I'm going to go about taking the user input and making a corresponding move before I can make these conditions.